Create private drivers for Input and Pagination, move things around in IframesManager
