+++ packages/wix-ui-core/src/components/Pagination/Pagination.spec.tsx
@@ -3,67 +3,11 @@ import * as ReactDOM from 'react-dom';
 import {Simulate} from 'react-dom/test-utils';
 import {StylableDOMUtil} from 'stylable/test-utils';
 import {sleep} from 'wix-ui-test-utils/react-helpers';
-import {Pagination} from '.';
+import {Pagination} from './Pagination';
+import {PaginationDriver} from './Pagination.private.driver';
 import style from './Pagination.st.css';
 import testStyle from './PaginationTest.st.css';
 import {ReactDOMTestContainer} from '../../../test/dom-test-container';
-
-class PaginationDriver {
-  constructor(public root: HTMLElement) { }
-
-  get previousButton() {
-    return this.root.querySelector('[data-hook=previous]');
-  }
-
-  get nextButton() {
-    return this.root.querySelector('[data-hook=next]');
-  }
-
-  get firstButton() {
-    return this.root.querySelector('[data-hook=first]');
-  }
-
-  get lastButton() {
-    return this.root.querySelector('[data-hook=last]');
-  }
-
-  get input(): HTMLInputElement {
-    return this.root.querySelector('[data-hook=page-input]');
-  }
-
-  get totalPagesLabel() {
-    return this.root.querySelector('[data-hook=total-pages]');
-  }
-
-  get pageStrip() {
-    return this.root.querySelector('[data-hook=page-strip]');
-  }
-
-  get pages() {
-    return Array.from(this.pageStrip.firstElementChild.children);
-  }
-
-  get pageLabels() {
-    return this.pages.map(p => p.textContent);
-  }
-
-  get currentPage() {
-    return this.root.querySelector('[data-hook~=current-page]');
-  }
-
-  getPage(n) {
-    return this.root.querySelector(`[data-hook~=page-${n}]`);
-  }
-
-  changeInput(value) {
-    this.input.value = value;
-    Simulate.change(this.input);
-  }
-
-  commitInput() {
-    Simulate.keyDown(this.input, {keyCode: 13});
-  }
-}
 
 const stylableUtil = new StylableDOMUtil(style);
